Create Cluster_sklearn.py
This script should load a .VCF file, extract the genotype data for each sample, convert the genotypes to numerical values, calculate a pairwise distance matrix between all samples based on the genotype data, perform clustering analysis using scikit-learn's KMeans algorithm with a specified number of clusters, and plot the results using matplotlib. Note that the script is just a simple example and may need to be modified depending on the specifics of your data and analysis.